sql/requirements.txt
matplotlib
pyowm
psycopg2-binary
streamlit
typer